Update index.html
In this example, we added an input element of type range to create a slider. We set the min and max attributes to specify the range of values for the slider. We also added an id attribute to the input element so that we can reference it in our JavaScript code.

In the script section, we added JavaScript code to listen for the input event on the colorSlider element. When the user moves the slider, we update the backgroundColor property of the body element using the hue value from the slider. We use an HSL (hue, saturation, lightness) color value to specify the background color.

--by bing